Rewrite using jQuery special events system
Rewrote the entire plugin to leverage jQuery's built-in special events system, which supports adding hooks for custom events (in this case, 'domchange'). This eliminates the need to call domready() on elements before listening to them. It also now supports having multiple listeners with different MutationObserver options on the same element (i.e. handlers with different sets of configuration options). Although potentially more performance intensive, this version is more user-friendly and is more customizable.